% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcovid19Canada.R
\name{runTestingPackage}
\alias{runTestingPackage}
\title{Launch Shiny App for covid19Canada}
\usage{
runTestingPackage()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for covid19Canada.
The purpose of this app is only to illustrate how a Shiny
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

covid19Canada::runcovid19Canada()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
