% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccinationCanada.R
\name{VaccinationCanada}
\alias{VaccinationCanada}
\title{Canada COVID-19 Vaccination Count}
\usage{
VaccinationCanada(dateOfInterest)
}
\arguments{
\item{dateOfInterest}{The string indicates the date of interest.}
}
\value{
Returns a data frame with 16 different columns that each indicates COVID-19 vaccination related counts.
}
\description{
A function that demonstrates the COVID-19 vaccination data in Canada. User
is able to enter the date of interested to look up the vaccination data of
Canada on that specific date.
}
\examples{
# Example 1:
#Use the function without arguments.

covidVaccination <- VaccinationCanada()
covidVaccination

# Example 2:
#Use the function only with the one argument.

covidVaccination <- VaccinationCanada("2021-01-01")
covidVaccination

}
\references{
Ponce et al. (2021). covid19.analytics: An R Package to Obtain, Analyze and
Visualize Data from the Coronavirus Disease Pandemic. Journal of Open Source
Software, 6(59), 2995. https://doi.org/10.21105/joss.02995
}
