% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfeVacPlot.R
\name{InfeVacPlot}
\alias{InfeVacPlot}
\title{Vaccination and Infection Plotting}
\usage{
InfeVacPlot()
}
\value{
Returns a plotting that demonstrates the correlation of
COVID-19 vaccination(vaccinated & fully vaccinated)and infection in Canada.
}
\description{
A function that generates the plotting demonstrating the correlation of
COVID-19 vaccination and infection in Canada.
}
\examples{
InfeVacPlot()

}
\references{
Ponce et al. (2021). covid19.analytics: An R Package to Obtain, Analyze and
Visualize Data from the Coronavirus Disease Pandemic. Journal of Open Source
Software, 6(59), 2995. https://doi.org/10.21105/joss.02995

R Core Team (2021). R: A language and
environment for statistical computing. R
Foundation for Statistical Computing, Vienna,
Austria. URL https://www.R-project.org/.
}
